% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/case_cut.R
\name{case_cut}
\alias{case_cut}
\title{Create a case statement by using like the base::cut().}
\usage{
case_cut(.data, ..., breaks, labels = NULL, include.lowest = FALSE,
  right = TRUE, dig.lab = 3)
}
\arguments{
\item{.data}{A tbl.}

\item{...}{a name-value pair of expression.}

\item{breaks}{a numeric vector of two or more unique cut points.}

\item{labels}{labels for the levels of the resulting category. By default, labels are constructed using "(a,b]" interval notation.}

\item{include.lowest}{logical, indicating if the lowest category contains the lowest value.}

\item{right}{logical, indicating if the intervals should be closed on the right (and open the left) or vice versa.}

\item{dig.lab}{integer which is used when labels are not given. It determines the number of digits used in formatting the break numbers.}
}
\description{
Create a case statement by using like the base::cut().
}

