% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mutate_first_value.R
\name{extract_first_value_info}
\alias{extract_first_value_info}
\title{extract first_value() arguments}
\usage{
extract_first_value_info(dot, d)
}
\arguments{
\item{dot}{a lazy befere the translation}

\item{d}{a lazy after the translation}
}
\value{
a lazy with arguments
}
\description{
extract first_value() arguments
}

