% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mutate_moving_something.R
\name{extract_moving_info}
\alias{extract_moving_info}
\title{extract moving_** arguments}
\usage{
extract_moving_info(d)
}
\arguments{
\item{d}{a lazy}
}
\value{
a lazy with arguments
}
\description{
extract moving_** arguments
}

