% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_.tbl_sql}
\alias{filter_.tbl_sql}
\title{Improved filter for tbl_sql which adds parentheses}
\usage{
\method{filter_}{tbl_sql}(.data, ..., .dots)
}
\arguments{
\item{.data}{A tbl_sql.}

\item{...}{Logical predicates. Multiple conditions are combined with &.}

\item{.dots}{Used to work around non-standard evaluation.}
}
\value{
An object of tbl_sql.
}
\description{
Improved filter for tbl_sql which adds parentheses
}

