% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/load_tbls.R
\name{load_tbls}
\alias{load_tbls}
\title{Load the table objects for all tables in a database}
\usage{
load_tbls(db, envir = parent.frame(), verbose = TRUE)
}
\arguments{
\item{db}{a database object}

\item{envir}{the environment to load table objects}
}
\description{
Load the table objects for all tables in a database
}

