% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/moving_.R
\name{moving_}
\alias{moving_}
\title{Compute moving ** for PostgreSQL}
\usage{
moving_(type = c("mean", "max", "min", "sum"), .data, ...)
}
\arguments{
\item{type}{A string.}

\item{.data}{A tbl_postgres.}

\item{...}{Name-value pairs of expressions.}
}
\value{
An object of the same class as .data.
}
\description{
Compute moving ** for PostgreSQL
}

