% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/moving_average.R
\name{moving_average}
\alias{moving_average}
\title{Compute moving average for PostgreSQL}
\usage{
moving_average(.data, ...)
}
\arguments{
\item{.data}{A tbl_postgres.}

\item{...}{Name-value pairs of expressions.}
}
\value{
An object of the same class as .data.
}
\description{
Compute moving average for PostgreSQL
}

