% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mutate.R
\name{mutate_.tbl_sql}
\alias{mutate_.tbl_sql}
\title{dplyr mutate() added original functions for databases}
\usage{
\method{mutate_}{tbl_sql}(.data, ..., .dots)
}
\description{
dplyr mutate() added original functions for databases
}

