% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mutate_count_if.R
\name{mutate_count_if}
\alias{mutate_count_if}
\title{translater for count_if() or n_if() added to mutate()}
\usage{
mutate_count_if(d)
}
\arguments{
\item{d}{a lazy before the translation}
}
\value{
a lazy after the translation
}
\description{
translater for count_if() or n_if() added to mutate()
}

