% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mutate_cut.R
\name{mutate_cut}
\alias{mutate_cut}
\title{translater for cut() added to mutate()}
\usage{
mutate_cut(d)
}
\arguments{
\item{d}{a lazy before the translation}
}
\value{
a lazy after the translation
}
\description{
translater for cut() added to mutate()
}

