% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mutate_first_value.R
\name{mutate_first_value}
\alias{mutate_first_value}
\title{translater for first_value() added to mutate()}
\usage{
mutate_first_value(d)
}
\arguments{
\item{d}{a lazy before the translation}
}
\value{
a lazy after the translation
}
\description{
translater for first_value() added to mutate()
}

