% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mutate_moving_something.R
\name{mutate_moving_something}
\alias{mutate_moving_something}
\title{translater for moving_**() added to mutate()}
\usage{
mutate_moving_something(type = c("mean", "max", "min", "sum"), d)
}
\arguments{
\item{d}{a lazy before the translation}
}
\value{
a lazy after the translation
}
\description{
translater for moving_**() added to mutate()
}

