% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_.tbl_sql}
\alias{filter_.tbl_sql}
\title{Improved filter for tbl_sql which adds parentheses}
\usage{
\method{filter_}{tbl_sql}(.data, ..., .dots)
}
\description{
Improved filter for tbl_sql which adds parentheses
}

