% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mutate.R
\name{mutate_.tbl_sql}
\alias{mutate_.tbl_sql}
\title{dplyr mutate() added original functions for databases}
\usage{
\method{mutate_}{tbl_sql}(.data, ..., .dots)
}
\arguments{
\item{.data}{A tbl_sql.}

\item{...}{Name-value pairs of expressions. Use NULL to drop a variable.}

\item{.dots}{Used to work around non-standard evaluation.}
}
\value{
An object of tbl_sql.
}
\description{
dplyr mutate() added original functions for databases
}

